requirements.txt
undetected-chromedriver
kafka-python
pathway